+++ sodalite/core/key.py
@@ -54,7 +54,7 @@ def get_all_keys():
     return keys
 
 
-def get_available_keys(old_entries: dict) -> List[str]:
+def _get_available_keys(old_entries: dict) -> List[str]:
     """returns a list of keys which are not yet used by given entries
     :param old_entries: list of Entry which already have a key"""
     used_keys = set(map(lambda x: x.key.value, old_entries.values()))
@@ -67,22 +67,42 @@ def get_available_keys(old_entries: dict) -> List[str]:
     return unused_keys
 
 
-def assign_keys(new_entries: dict, old_entries: dict):
+def assign_keys(entries_to_assign: dict, old_entries: dict):
     """ assigns keys to the given new entries. Needs old entries
-    :param new_entries: entries without key, these will receive a key
-    :param old_entries: all entries of this domain which already have a key"""
-    free_keys = get_available_keys(old_entries)
-    sorted_new_entries = _sort(new_entries.values())
+    :param entries_to_assign: entries without key, these will receive a key
+    :param old_entries: all entries of this domain which already have a key
+    :return list of old entries whose keys got reassigned and therefore need to get persisted
+    """
+
+    if not entries_to_assign:
+        return
+    free_keys = _get_available_keys(old_entries)
+    reassignable_keys = [x.key.value for x in old_entries.values() if x.rating < 0.05 and x.key.value != '']
+    available_keys = free_keys + reassignable_keys
+    sorted_new_entries = _sort(entries_to_assign.values())
+    _assign(sorted_new_entries, available_keys)
+
+    # now try reassign keys to these entries who just lost their key
+    reassigned_keys = [x for x in reassignable_keys if x not in available_keys]
+    entries_to_reassign = [x for x in old_entries.values() if x.key.value in reassigned_keys]
+    for entry in entries_to_reassign:
+        entry.key = Key('')
+    _assign(entries_to_reassign,
+            [x for x in free_keys if x not in available_keys])
+    return entries_to_reassign
+
+
+def _assign(entries: List, available_keys: List[str]):
     entries_assign_later = []
-    for entry in sorted_new_entries:
-        if len(free_keys) > 0:
+    for entry in entries:
+        if len(available_keys) > 0:
             # try to assign starting character as key
             start_char = entry.name[0].lower()
-            if start_char in free_keys:
-                free_keys.remove(start_char)
+            if start_char in available_keys:
+                available_keys.remove(start_char)
                 char = start_char
-            elif start_char.upper() in free_keys:
-                free_keys.remove(start_char.upper())
+            elif start_char.upper() in available_keys:
+                available_keys.remove(start_char.upper())
                 char = start_char.upper()
             else:
                 entries_assign_later.append(entry)
@@ -92,8 +112,8 @@ def assign_keys(new_entries: dict, old_entries: dict):
             return
     # now assign keys for entries whose starting chars were already taken
     for entry in entries_assign_later:
-        if len(free_keys) > 0:
-            char = free_keys.pop(0)
+        if len(available_keys) > 0:
+            char = available_keys.pop(0)
             entry.key = Key(char)
         else:
             return